requirements.txt
aioconsole==0.8.0
openmeteo_requests==1.3.0
pandas==2.2.3
python-dotenv==1.0.1
requests_cache==1.2.1
retry_requests==2.0.0
SQLAlchemy==2.0.35
